+++ src/main/kotlin/systems/fehn/intellijdirenv/services/DirenvProjectService.kt
@@ -16,6 +16,7 @@ import com.intellij.openapi.project.guessProjectDir
 import com.intellij.openapi.vfs.VirtualFile
 import systems.fehn.intellijdirenv.MyBundle
 import systems.fehn.intellijdirenv.notificationGroup
+import systems.fehn.intellijdirenv.settings.DirenvSettingsState
 import systems.fehn.intellijdirenv.switchNull
 
 class DirenvProjectService(private val project: Project) {
@@ -69,7 +70,6 @@ class DirenvProjectService(private val project: Project) {
                         MyBundle.message("exceptionNotification"),
                         e.localizedMessage,
                         NotificationType.ERROR,
-                        null,
                     )
             }
         }
@@ -140,8 +140,14 @@ class DirenvProjectService(private val project: Project) {
     private fun executeDirenv(envrcFile: VirtualFile, vararg args: String): Process {
         val workingDir = envrcFile.parent.path
 
-        return GeneralCommandLine("direnv", *args)
+        val cli = GeneralCommandLine("direnv", *args)
             .withWorkDirectory(workingDir)
-            .createProcess()
+
+        val appSettings = DirenvSettingsState.getInstance()
+        if (appSettings.direnvSettingsPath.isNotEmpty()) {
+            cli.withExePath(appSettings.direnvSettingsPath)
+        }
+
+        return cli.createProcess()
     }
 }